backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.4
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
requests==2.31.0
httpx==0.25.2
python-decouple==3.8
gunicorn==21.2.0
django-extensions==3.2.3
openpyxl==3.1.2
django-celery-beat==2.5.0
whitenoise==6.6.0